requirements.txt
python-telegram-bot[asyncio]~=20.6
aiohttp
telethon
python-dotenv